refactor: extract Sparkle updater into UpdaterService
